+++ ecosystem/indexer-grpc/indexer-grpc-utils/src/lib.rs
@@ -8,10 +8,12 @@ pub mod file_store_operator;
 pub mod types;
 
 use aptos_protos::{
+    indexer::v1::raw_data_client::RawDataClient,
     internal::fullnode::v1::fullnode_data_client::FullnodeDataClient, transaction::v1::Transaction,
     util::timestamp::Timestamp,
 };
 use prost::Message;
+use std::time::Duration;
 use url::Url;
 
 pub type GrpcClientType = FullnodeDataClient<tonic::transport::Channel>;
@@ -31,6 +33,41 @@ pub async fn create_grpc_client(address: Url) -> GrpcClientType {
                 tracing::error!(
                     address = address.to_string(),
                     "[Indexer Cache] Failed to connect to indexer gRPC server: {}",
+                    e
+                );
+                Err(backoff::Error::transient(e))
+            },
+        }
+    })
+    .await
+    .unwrap()
+}
+
+pub type GrpcDataServiceClientType = RawDataClient<tonic::transport::Channel>;
+
+/// Create a gRPC client for the indexer data service with exponential backoff.
+/// max_elapsed_time is the maximum time to wait for the connection to be established.
+pub async fn create_data_service_grpc_client(
+    address: Url,
+    max_elapsed_time: Option<Duration>,
+) -> GrpcDataServiceClientType {
+    let mut backoff = backoff::ExponentialBackoff::default();
+    if let Some(max_elapsed_time) = max_elapsed_time {
+        backoff.max_elapsed_time = Some(max_elapsed_time);
+    }
+    backoff::future::retry(backoff, || async {
+        match RawDataClient::connect(address.to_string()).await {
+            Ok(client) => {
+                tracing::info!(
+                    address = address.to_string(),
+                    "[Indexer Cache] Connected to indexer data service gRPC server."
+                );
+                Ok(client)
+            },
+            Err(e) => {
+                tracing::error!(
+                    address = address.to_string(),
+                    "[Indexer Cache] Failed to connect to indexer data service gRPC server: {}",
                     e
                 );
                 Err(backoff::Error::transient(e))